refactor: ThemeChooser 组件js改为隔离
