+++ src/BootstrapBlazor.Shared/Components/ThemeChooser.razor
@@ -1,5 +1,5 @@
 ﻿@inherits WebSiteModuleComponentBase
-@attribute [JSModuleAutoLoader("theme-chooser")]
+@attribute [JSModuleAutoLoader("Components/ThemeChooser.razor.js")]
 
 <div id="@Id" class="theme">
     <PulseButton class="btn-fade btn-theme" Color="Color.None" ImageUrl="_content/BootstrapBlazor.Shared/images/m.svg" TooltipText="@Title" TooltipPlacement="Placement.Left" />
